Isolating createBitcoinMultisig function
